requirements.txt
django == 1.11
djangorestframework == 3.8.2
psycopg2-binary
django-computed-property